// frontend/src/openapi/models/FundingSourceEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `platform` - platform
 * * `chatgpt` - chatgpt
 * @export
 */
export const FundingSourceEnum = {
    platform: 'platform',
    chatgpt: 'chatgpt'
} as const;
export type FundingSourceEnum = typeof FundingSourceEnum[keyof typeof FundingSourceEnum];


export function instanceOfFundingSourceEnum(value: any): boolean {
    for (const key in FundingSourceEnum) {
        if (Object.prototype.hasOwnProperty.call(FundingSourceEnum, key)) {
            if (FundingSourceEnum[key as keyof typeof FundingSourceEnum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function FundingSourceEnumFromJSON(json: any): FundingSourceEnum {
    return FundingSourceEnumFromJSONTyped(json, false);
}

export function FundingSourceEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): FundingSourceEnum {
    return json as FundingSourceEnum;
}

export function FundingSourceEnumToJSON(value?: FundingSourceEnum | null): any {
    return value as any;
}

export function FundingSourceEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): FundingSourceEnum {
    return value as FundingSourceEnum;
}